@@ -6,7 +6,8 @@ import { BrowserRouter as Router, Route,Routes } from 'react-router-dom';
 import Customer from './components/Customer/Customer';
 import Header from './components/Header/Header';
 import TermDeposit from './components/Deposits/Term-deposit';
-import Chairman_message from './components/About_Us/Chairman_message/chairman_message';
+import Branch from './components/Branch/Branch';
+// import Chairmanmessage from './components/About_Us/Chairman_message/chairman_message';
 // import AddCustomer from './components/Customer/AddCustomer';
 
 function App() {
@@ -20,7 +21,8 @@ function App() {
         <Route path="/home" element={<Home />} />
         <Route path="/customer" element={<Customer />} />
         <Route path="/term-deposit" element={<TermDeposit />} />
-        <Route path="/chairman-message" element={<Chairman_message/>} />
+        <Route path="/branches" element={<Branch />} />
+        {/* <Route path="/chairman-message" element={<Chairmanmessage/>} /> */}
         <Route path="/" element={<Home />} />
       </Routes>
     </div>
