@@ -23,20 +23,21 @@ function Header() {
           </a>
         </div>
         <div className="rightNavBar" style={{margin:'20px',fontFamily:'Reem Kufi Ink,sans-serif'}}>
-          <ul className="navbar-nav me-auto mb-2 mb-lg-0" style={{display: 'flex',flexDirection: 'row',alignItems: 'center',gap:'15px',fontFamily:'cursive'}}>
+          <ul className="navbar-nav me-auto mb-2 mb-lg-0" style={{display: 'flex',flexDirection: 'row',alignItems: 'center',gap:'15px',fontFamily:'sans-serif',fontWeight:'bolder'}}>
            { 
              
               headerJson.menu.map((element)=>{
+               
                 return(
                   <li className="nav-item dropdown header-tab-name" key={element.menu}>
-                      <a className={element.submenu.length>0?'nav-link dropdown-toggle':'nav-link'}   href="/"  id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                      <a className={element.submenu.length>0?'nav-link dropdown-toggle':'nav-link'}  alt='' href={element.menuUrl}  id="navbarDropdown" role="button" data-bs-toggle={element.submenu.length>0?'dropdown':''} aria-expanded="false">
                           {element.menu}
                       </a>
                       {
-                        element.submenu.length>0 && 
+                        element.submenu?.length>0 && 
                             <ul className="dropdown-menu" aria-labelledby="navbarDropdown" style={{boxShadow:'inset 1px 2px 14px 8px #507d0d47',backgroundColor:'#fefff5'}} key={element.menu+'1'}>
                             {
-                              element.submenu.map(submenu => {
+                              element.submenu?.map(submenu => {
                                 return(
                                 <li key={submenu.name} style={{boxShadow: '1px 2px 13px 1px #507d0d30',padding:'7px'}}><a className="dropdown-item" href={submenu.url}>{submenu.name}</a></li>
                                 
